% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS csv data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A csv path}
}
\value{
A data frame tbl object.
}
\description{
This function reads FARS (Fatality Analysis Reporting System) data from a csv input and returns a
data frame tbl (thin wrapper around data.frame) object
}
\details{
If the input file does not exists, an error will be thrown with the message
    "file <requested_file> does not exists"
}
\examples{
\dontrun{
  fars_read("accident_2013.csv.bz2")
}
}
