% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read the available year, month from the dataset}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The years to be read}
}
\value{
A list of data frames corresponding to the requested years. Each data frame contains
    column: MONTH and year, indicating the avaiable year-month from the data
}
\description{
This function reads the available month, year combinations from the dataset
}
\details{
If a year results in error (file not available, corrupted data, etc) a warning will be
    shown and a NULL object is returned
}
\examples{
  fars_read_years(c(2013,2014))

}
