% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the data for given years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The years to be read}
}
\value{
A data frame of one row, and as many columns as the requested years, the cell represents
    the number of records for the given year
}
\description{
This function takes some years as input and return a dataframe showing the number of records per
each year
}
\examples{
\dontrun{
  fars_summarize_years(c(2013,2014)
}

}
